Storage of memory body text moved from db to files
Database now containers only each memory's title and keywords in order to perform string matching. The body text for each memory is now in a separate data directory. Body content is retrieved by accessing the file whose name corresponds to the db_key of the database record for that memory.