Made update functions more generic, narrowed mode switching to DataController, enabled scrolling, implemented saving changes to linked content when embedded instruction used
Rather than specializing the update functions to update several hardcoded variables, I made it so that they only updated the variables that have a matching key in the input parameter.

I also began encapsulating manipulation of mode variables to the DataController class to prevent problems later in determining where the mode was last changed.

If a record's interpretted body text is altered, that change is now reflected in the linked file. Plan to change embedded instruction implementation later so that any change to a linked piece of content gets changed in that contents source, even if multiple exist in the same record.