Fixed keyword widget bug and made slight improvements to prevent similar bugs
Keyword widget was not being show because it was being remove the the body listbox as a result of the UI starting in COMMAND mode. It was removing it although it was not present to begin with.

UI initialization now performed using _set_interaction_mode and _set_view_mode instead of passing modes as constructor arguments.

Determining whether a record needs to be saved is now done by taking a hash of the content when it is initialized and comparing it to a hash of the content at exit. This solved a bug where a field started out empty, but was changed and then made empty again. Prior implementation was seeing this as a change and indicating that updates to the database were required.