Record changes now accurately saved on close
Any non-interpreted change that alters a record is updated in the database. If a change is made to interpreted PATH text and the raw instruction is also altered, the linked PATH file will reflect the changes and the updated raw instruction will also change in the database although the prior PATH link will no longer exist.