Bug fix - Database now looked for in user's home directory
Bug was that the program would look for the datadir directory in the current working directory and create it if not found. This meant that the user would always need to run the program from the same directory in order to be using the same database. Fixed so that folder is created in the user's home directory to contain the record database.